test(core): move DecodedFrame and AVC helper tests into testsuite-core
ironrdp-egfx sets [lib] test = false commented out with a FIXME to
turn it back on once its tests move into the testsuite crates. Two of
its five inline test files test only public API and can move cleanly:
DecodedFrame's constructor and getters in decode.rs, and
Avc420Region/align_to_16/annex_b_to_avc/encode_avc420_bitmap_stream in
pdu/avc.rs. Moved both, verbatim apart from a couple of renamed test
functions: this directory's existing files split between a test_
prefix (decode.rs, server.rs) and plain descriptive names
(capabilities.rs, client.rs), and the moved tests follow the latter.

DecodedFrame's tests go in a new, unconditional decoded_frame.rs: the
type itself needs no feature flag, unlike the existing decode.rs test
file which is gated on openh264-bundled for the OpenH264Decoder tests
it already covers.

The remaining three files (client.rs, server.rs, compositor.rs) test
private internals and stay inline; this does not change the crate's
test = false state, which is unrelated to this move and stays as it
was.
