+++ crates/ironrdp-testsuite-core/tests/egfx/mod.rs
@@ -1,5 +1,7 @@
+mod avc;
 mod capabilities;
 mod client;
 #[cfg(feature = "openh264-bundled")]
 mod decode;
+mod decoded_frame;
 mod server;
